DiseaseIdentification/backend/backend/requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
flask==3.0.2
fonttools==4.49.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.2
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
pillow==10.2.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
thop==0.1.1.post2209072238
torchvision==0.17.1
tqdm==4.66.2
typing-extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.19
urllib3==2.2.1
werkzeug==3.0.1
zipp==3.17.0
